+++ src/omni/base.py
@@ -52,6 +52,7 @@ from . import transfer
 from . import document
 from . import employee
 from . import sale_order
+from . import export_api
 from . import credit_note
 from . import sub_product
 from . import merchandise
@@ -119,6 +120,7 @@ class API(
     document.DocumentAPI,
     employee.EmployeeAPI,
     sale_order.SaleOrderAPI,
+    export_api.ExportAPI,
     credit_note.CreditNoteAPI,
     sub_product.SubProductAPI,
     merchandise.MerchandiseAPI,